ComprehensionScience/requirements.txt
streamlit==1.13.0
streamlit-option-menu==0.3.2
sumy==0.10.0
certifi==2022.9.24
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
beautifulsoup4==4.11.1
requests==2.28.1
requests-toolbelt>=0.9.1
htbuilder==0.6.0
pandas==1.5.0
Pillow>=7.1.2
numpy==1.23.4
nltk==3.7
urllib3==1.26.12
